pyproject.toml
[tool.poetry]
name = "chat"
version = "0.1.0"
description = ""
authors = ["Joel Zamboni <joel.zamboni@webera.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
langchain = "^0.1.17"
chainlit = "^1.0.506"
langchainhub = "^0.1.15"
chromadb = "^0.5.0"
pypdf = "^4.2.0"
tldextract = "^5.1.2"
PyPDF2 = "^3.0.1"
spacy = "^3.7.4"
transformers = "^4.40.2"
async-timeout = "^4.0.3"
beautifulsoup4 = "^4.12.3"
requests-html = "^0.10.0"
lxml-html-clean = "^0.1.1"
websockets = "^12.0"
selenium = "^4.20.0"
webdriver-manager = "^4.0.1"
speechrecognition = "^3.10.4"
pyttsx3 = "^2.90"
pydub = "^0.25.1"
gtts = "^2.5.1"
setuptools-rust = "^1.9.0"
openai-whisper = { git = "https://github.com/openai/whisper.git" }

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
